# examples/petstore/generated/dto/dto.go
// Code generated by apiggo. DO NOT EDIT.
package dto

type Category struct {
	Id   *int64  `json:"id,omitempty"`
	Name *string `json:"name,omitempty"`
}

type CreatePetBadRequest struct {
	Code    *int32  `json:"code,omitempty"`
	Message *string `json:"message,omitempty"`
}

func (e *CreatePetBadRequest) Error() string   { return "Bad Request" }
func (e *CreatePetBadRequest) StatusCode() int { return 400 }
func (e *CreatePetBadRequest) Body() any       { return e }

type CreatePetIn struct {
	Body PetCreate `json:"-"`
}

type CreatePetOut = Pet

type GetPetIn struct {
	Id      int64 `json:"id"`
	Verbose bool  `json:"verbose"`
}

type GetPetNotFound struct {
	Message *string `json:"message,omitempty"`
}

func (e *GetPetNotFound) Error() string   { return "Not Found" }
func (e *GetPetNotFound) StatusCode() int { return 404 }
func (e *GetPetNotFound) Body() any       { return e }

type GetPetOut = Pet

type Pet struct {
	Category *Category `json:"category,omitempty"`
	Id       int64     `json:"id"`
	Name     string    `json:"name"`
	Status   PetStatus `json:"status"`
	Tags     []string  `json:"tags,omitempty"`
}

type PetCreate struct {
	Name   string           `json:"name"`
	Status *PetCreateStatus `json:"status,omitempty"`
}

type PetCreateStatus string

const (
	PetCreateStatusAvailable PetCreateStatus = "available"
	PetCreateStatusPending   PetCreateStatus = "pending"
	PetCreateStatusSold      PetCreateStatus = "sold"
)

type PetStatus string

const (
	PetStatusAvailable PetStatus = "available"
	PetStatusPending   PetStatus = "pending"
	PetStatusSold      PetStatus = "sold"
)