# examples/petstore/generated/router/router.go
// Code generated by apiggo. DO NOT EDIT.
package router

import (
	"context"
	"encoding/json"
	"net/http"
	"strconv"

	dto "github.com/siyoga/apiggo/examples/petstore/generated/dto"
	"github.com/siyoga/apiggo/server"
)

// CreatePet registers the POST /pets operation.
func CreatePet(h func(ctx context.Context, in dto.CreatePetIn) (dto.CreatePetOut, error)) func(*server.Server) {
	return func(s *server.Server) {
		s.Register(http.MethodPost, "/pets", http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
			var in dto.CreatePetIn
			if err := json.NewDecoder(r.Body).Decode(&in.Body); err != nil {
				s.HandleError(w, r, server.ErrBadRequest("invalid request body"))
				return
			}
			out, err := h(r.Context(), in)
			if err != nil {
				s.HandleError(w, r, err)
				return
			}
			w.Header().Set("Content-Type", "application/json")
			_ = json.NewEncoder(w).Encode(out)
		}))
	}
}

// GetPet registers the GET /pets/{id} operation.
func GetPet(h func(ctx context.Context, in dto.GetPetIn) (dto.GetPetOut, error)) func(*server.Server) {
	return func(s *server.Server) {
		s.Register(http.MethodGet, "/pets/{id}", http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
			var in dto.GetPetIn
			{
				raw := r.PathValue("id")
				v, err := strconv.ParseInt(raw, 10, 64)
				if err != nil {
					s.HandleError(w, r, server.ErrBadRequest("invalid parameter: id"))
					return
				}
				in.Id = v
			}
			if raw := r.URL.Query().Get("verbose"); raw != "" {
				v, err := strconv.ParseBool(raw)
				if err != nil {
					s.HandleError(w, r, server.ErrBadRequest("invalid parameter: verbose"))
					return
				}
				in.Verbose = v
			}
			out, err := h(r.Context(), in)
			if err != nil {
				s.HandleError(w, r, err)
				return
			}
			w.Header().Set("Content-Type", "application/json")
			_ = json.NewEncoder(w).Encode(out)
		}))
	}
}