# testdata/petstore-min/openapi.yaml
openapi: 3.0.3
info:
  title: petstore-min
  version: 1.0.0
paths:
  /pets/{id}:
    get:
      operationId: getPet
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
        - name: verbose
          in: query
          schema: { type: boolean }
        - name: X-Trace
          in: header
          schema: { type: string }
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Pet' }
        '404':
          description: not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
  /pets:
    post:
      operationId: createPet
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PetCreate' }
      responses:
        '200':
          description: created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Pet' }
        '400':
          description: bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  code: { type: integer, format: int32 }
components:
  schemas:
    Pet:
      type: object
      required: [id, name, status]
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        status: { type: string, enum: [available, pending, sold] }
        tags:
          type: array
          items: { type: string }
        category:
          $ref: '#/components/schemas/Category'
    Category:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
    PetCreate:
      type: object
      required: [name]
      properties:
        name: { type: string }
        status: { type: string, enum: [available, pending, sold] }